requirements.txt
PyMuPDF
pandas
spacy
scikit-learn
pdf2image
Pillow
pytesseract
streamlit